real-time-student-support-system-backend-main/requirements.txt
Django==5.2

djangorestframework==3.16.1
django-redis==6.0.0
django-environ==0.12.0

django-cors-headers==4.9.0

# uvicorn==0.40.0
uvicorn[standard]==0.40.0

channels==4.3.2
channels-redis==4.3.0

whitenoise==6.11.0

psycopg2==2.9.11

celery==5.6.0
django-celery-results==2.6.0
django-celery-beat==2.8.1

djangorestframework-simplejwt==5.5.1

# strawberry-graphql==0.287.3
strawberry-graphql[channels]==0.287.3
strawberry-graphql-django==0.70.1

